2.1) Начинаем программировать Устранение замечания наставника Переименование компонентов относительно названий с разметки
