+++ project/src/components/favorites-empty/favorites-empty.tsx
@@ -1,4 +1,4 @@
-function FavoritesPageEmpty(): JSX.Element {
+function FavoritesEmpty(): JSX.Element {
   return (
     <div className="page page--favorites-empty">
       <header className="header">
@@ -49,4 +49,4 @@ function FavoritesPageEmpty(): JSX.Element {
   );
 }
 
-export default FavoritesPageEmpty;
+export default FavoritesEmpty;
