+++ project/src/components/favorites/favorites.tsx
@@ -1,4 +1,4 @@
-function FavoritesPage(): JSX.Element {
+function Favorites(): JSX.Element {
   return (
     <div className="page">
       <header className="header">
@@ -164,4 +164,4 @@ function FavoritesPage(): JSX.Element {
   );
 }
 
-export default FavoritesPage;
+export default Favorites;
