+++ project/src/components/main/main.tsx
@@ -1,4 +1,10 @@
-function MainPageEmpty(): JSX.Element {
+import PlaceCard from '../place-card/place-card';
+
+type MainPageProps = {
+  placeCardCount: number;
+}
+
+function Main({placeCardCount}: MainPageProps): JSX.Element {
   return (
     <div className="page page--gray page--main">
       <header className="header">
@@ -29,7 +35,7 @@ function MainPageEmpty(): JSX.Element {
         </div>
       </header>
 
-      <main className="page__main page__main--index page__main--index-empty">
+      <main className="page__main page__main--index">
         <h1 className="visually-hidden">Cities</h1>
         <div className="tabs">
           <section className="locations container">
@@ -50,7 +56,7 @@ function MainPageEmpty(): JSX.Element {
                 </a>
               </li>
               <li className="locations__item">
-                <a className="locations__item-link tabs__item" href="/">
+                <a className="locations__item-link tabs__item tabs__item--active" href="/">
                   <span>Amsterdam</span>
                 </a>
               </li>
@@ -60,7 +66,7 @@ function MainPageEmpty(): JSX.Element {
                 </a>
               </li>
               <li className="locations__item">
-                <a className="locations__item-link tabs__item tabs__item--active" href="/">
+                <a className="locations__item-link tabs__item" href="/">
                   <span>Dusseldorf</span>
                 </a>
               </li>
@@ -68,14 +74,37 @@ function MainPageEmpty(): JSX.Element {
           </section>
         </div>
         <div className="cities">
-          <div className="cities__places-container cities__places-container--empty container">
-            <section className="cities__no-places">
-              <div className="cities__status-wrapper tabs__content">
-                <b className="cities__status">No places to stay available</b>
-                <p className="cities__status-description">We could not find any property available at the moment in Dusseldorf</p>
+          <div className="cities__places-container container">
+            <section className="cities__places places">
+              <h2 className="visually-hidden">Places</h2>
+              <b className="places__found">{placeCardCount} places to stay in Amsterdam</b>
+              <form className="places__sorting" action="#" method="get">
+                <span className="places__sorting-caption">Sort by</span>
+                <span className="places__sorting-type" tabIndex={0}>
+                  Popular
+                  <svg className="places__sorting-arrow" width="7" height="4">
+                    <use xlinkHref="#icon-arrow-select"/>
+                  </svg>
+                </span>
+                <ul className="places__options places__options--custom places__options--opened">
+                  <li className="places__option places__option--active" tabIndex={0}>Popular</li>
+                  <li className="places__option" tabIndex={0}>Price: low to high</li>
+                  <li className="places__option" tabIndex={0}>Price: high to low</li>
+                  <li className="places__option" tabIndex={0}>Top rated first</li>
+                </ul>
+              </form>
+              <div className="cities__places-list places__list tabs__content">
+                <PlaceCard />
+                <PlaceCard />
+                <PlaceCard />
+                <PlaceCard />
+                <PlaceCard />
+                <PlaceCard />
               </div>
             </section>
-            <div className="cities__right-section"/>
+            <div className="cities__right-section">
+              <section className="cities__map map"/>
+            </div>
           </div>
         </div>
       </main>
@@ -83,4 +112,4 @@ function MainPageEmpty(): JSX.Element {
   );
 }
 
-export default MainPageEmpty;
+export default Main;
