+++ project/src/index.tsx
@@ -3,11 +3,11 @@ import ReactDOM from 'react-dom';
 import App from './components/app/app';
 
 const Param = {
-  POINT_CARD_COUNT: 6,
+  PLACE_CARD_COUNT: 6,
 };
 
 ReactDOM.render(
   <React.StrictMode>
-    <App pointCardCount={Param.POINT_CARD_COUNT}/>
+    <App placeCardCount={Param.PLACE_CARD_COUNT}/>
   </React.StrictMode>,
   document.getElementById('root'));
